+++ releasenotes.md
@@ -1,3 +1,4 @@
 Whats New in v11.2
 --------------------------
+- Add full usage tab for formula per IFRA category
 - This release may include more changes, for full details please refer to the CHANGELOG